+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.24.1
+
+- `execute-generators`: return generation failed status in case of any reported errors during the generation, 
+independent of the generation result, which might still be successful (consistent with the generator behavior in MPS 
+generated Ant scripts).
+
 ## 1.24.0
 
 - `modelcheck`: Include the model stereotype when matching (`--model` and `--exclude-model` parameters) and in the JUnit